fix: resolve 20 CodeQL alerts — overflow, dispose, null dereference
- DateShifter: cast weeks to double before multiplication (cs/loss-of-precision)
- ExecutionPlanPreviewDialog: store buttons as fields (cs/local-not-disposed)
- Handler tests: replace `as` casts with direct casts (cs/dereferenced-value-may-be-null)

+++ src/PPDS.Migration/Import/DateShifter.cs
@@ -33,7 +33,7 @@ namespace PPDS.Migration.Import
         private static TimeSpan RoundToWeeks(TimeSpan elapsed)
         {
             var weeks = (long)Math.Round(elapsed.TotalDays / 7.0);
-            return TimeSpan.FromDays(weeks * 7);
+            return TimeSpan.FromDays((double)weeks * 7);
         }
 
         private static TimeSpan RoundToDays(TimeSpan elapsed)
